inventory/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
black==20.8b1
click==7.1.2
Django==5.0
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.14.0
isort==5.7.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==10.1.0
pytz==2021.1
regex==2020.11.13
sqlparse==0.4.4
toml==0.10.2
typed_ast==1.5.5
typing-extensions==3.7.4.3
tzdata==2023.4
